Use Python 3.8+ and bump dependencies

+++ intro/matplotlib/examples/plot_imshow.py
@@ -12,8 +12,8 @@ def f(x, y):
     return (1 - x / 2 + x ** 5 + y ** 3 ) * np.exp(-x ** 2 - y ** 2)
 
 n = 10
-x = np.linspace(-3, 3, 3.5 * n)
-y = np.linspace(-3, 3, 3.0 * n)
+x = np.linspace(-3, 3, int(3.5 * n))
+y = np.linspace(-3, 3, int(3.0 * n))
 X, Y = np.meshgrid(x, y)
 Z = f(X, Y)
